Handle unavailable commands as commands that do not exist
If a participant used a command that was not available for her the raw
command was shown in the chat for everyone. Now the command is treated
like a command that does not exist (a "The command does not exist"
message is shown only to the user that called the command).

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>